+++ core/services/__init__.py
@@ -1,0 +1,12 @@
+"""core.services: Contains the core business logic and data processing for the ACE application.
+
+Services coordinate the application's workflows. They utilise Adapters to retrieve
+raw data from external sources, then parse, format, and structure that data so it
+can be easily consumed by the application's tools and models.
+"""
+
+from .database import SQLiteDatabaseService
+from .location import IPInfoLocationService
+from .weather import OpenWeatherMapService
+
+__all__ = ["OpenWeatherMapService", "IPInfoLocationService", "SQLiteDatabaseService"]
