+++ tests/services/__init__.py
@@ -1,0 +1,1 @@
+"""tests.services: Ensure that services are correctly defined and can be used by tools."""
